+++ cf/cfdatetime.py
@@ -166,8 +166,8 @@ def dt(arg, month=1, day=1, hour=0, minute=0, second=0,
 
 
 def dt_vector(arg, month=1, day=1, hour=0, minute=0, second=0,
-              microsecond=0, calendar=None):
-    '''Return a scalar or 1-d array of date-time objects.
+              microsecond=0, calendar=''):
+    '''Return a 1-d array of date-time objects.
 
     .. seealso:: `cf.dt`
 
@@ -200,7 +200,7 @@ def dt_vector(arg, month=1, day=1, hour=0, minute=0, second=0,
     :Returns:
 
         `numpy.ndarray`
-            The scalar or 1-d array of date-time objects.
+            1-d array of date-time objects.
 
     **Examples:**
 
@@ -235,7 +235,12 @@ def dt_vector(arg, month=1, day=1, hour=0, minute=0, second=0,
         if ndim >= 1:
             out = [out]
 
-        return numpy.array(out)
+        out = numpy.array(out)
+
+        if not out.ndim:
+            out = numpy.expand_dims(out, 0)
+
+        return out
 
     if arg.ndim == 2 and arg.shape[1] > 7:
         raise ValueError('TODO (890)')
@@ -260,11 +265,17 @@ def dt_vector(arg, month=1, day=1, hour=0, minute=0, second=0,
             _[:, 5] = second
             _[:, 6] = microsecond
             arg = _
-
+            
             out = [dt(*args, calendar=calendar) for args in arg]
-    # --- End: if
+    else:
+         out = [dt(*args, calendar=calendar) for args in arg]
 
-    return numpy.array(out)
+    out = numpy.array(out)
+
+    if not out.ndim:
+        out = numpy.expand_dims(out, 0)
+
+    return out
 
 
 def st2dt(array, units_in=None, dummy0=None, dummy1=None):