+++ cf/test/test_DomainAncillary.py
@@ -17,8 +17,8 @@ class DomainAncillaryTest(unittest.TestCase):
         _ = str(f)
         _ = f.dump(display=False)
 
-
 # --- End: class
+
 
 if __name__ == '__main__':
     print('Run date:', datetime.datetime.utcnow())